fix(tests): move the sync fake_stream tests into the mapped file
tests/test_litellm/llms/chat/test_converse_handler.py is already the mapped
test file for bedrock's converse_handler and already imports make_sync_call.
Adding a second file with the same basename under bedrock/chat broke
collection, since neither directory carries an __init__.py so pytest derives
the module name from the basename alone:

  import file mismatch: imported module 'test_converse_handler' has this
  __file__ attribute: .../llms/bedrock/chat/test_converse_handler.py
  which is not the same as the test file we want to collect:
  .../llms/chat/test_converse_handler.py

Extend the existing file instead and drop the duplicate.
